requirements.txt
# 视频去重工具 - Python依赖包

# 核心依赖
opencv-python>=4.5.0
Pillow>=8.0.0
imagehash>=4.2.0
send2trash>=1.8.0

# 可选依赖（用于性能优化）
numpy>=1.19.0

# 打包工具（仅开发环境需要）
pyinstaller>=6.0.0
